// client/api/models/ReservacionesListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReservacionesListItem = {
    id?: number;
    folio?: string | null;
    nombre?: string | null;
    correoElectronico?: string | null;
    telefono?: string | null;
    eventoId?: number;
    zonaEventoId?: number | null;
    estadoDeReservacionId?: number;
    cantidadBoletos?: number;
    precioUnitario?: number | null;
    subtotal?: number | null;
    fechaReservacion?: string;
    fechaExpiracion?: string;
    fechaPago?: string | null;
    fechaCreacion?: string;
    fechaModificacion?: string | null;
    estadoDeReservacion?: string | null;
    evento?: string | null;
    zonaEvento?: string | null;
};