// client/api/models/ZonasEventoListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ZonasEventoListItem = {
    id?: number;
    eventoId?: number;
    nombre?: string | null;
    capacidad?: number;
    precio?: number;
    fechaCreacion?: string;
    creadoPor?: string | null;
    fechaModificacion?: string | null;
    modificadoPor?: string | null;
    evento?: string | null;
};